Revert "refactor(destination-redshift-v2): replace RedshiftSqlGenerator with RedshiftTableOperationsClient"
This reverts commit 8b51058e98b296d16b201194490ebb0f72821ed2.
